Revert "Show social review prompt with both Social and Jetpack installed (#29910)"
This reverts commit 4f156469ee265d3fbf6bc2ecb283dfdd468046e9.
