--- projects/plugins/social/changelog/update-show-social-review-prompt-in-core
@@ -1,4 +1,0 @@
-Significance: minor
-Type: changed
-
-Moved review prompt into separate script
