+++ projects/plugins/social/src/js/editor.js
@@ -1,4 +1,4 @@
-import { JetpackLogo, SocialIcon } from '@automattic/jetpack-components';
+import { JetpackLogo, SocialIcon, getRedirectUrl } from '@automattic/jetpack-components';
 import {
 	SocialPreviewsModal,
 	SocialPreviewsPanel,
@@ -6,7 +6,11 @@ import {
 	usePublicizeConfig,
 	useSocialMediaConnections,
 	PublicizePanel,
+	ReviewPrompt,
+	usePostStartedPublishing,
 } from '@automattic/jetpack-publicize-components';
+import { getJetpackData } from '@automattic/jetpack-shared-extension-utils';
+import apiFetch from '@wordpress/api-fetch';
 import { PanelBody } from '@wordpress/components';
 import { dispatch, useSelect } from '@wordpress/data';
 import domReady from '@wordpress/dom-ready';
@@ -14,6 +18,7 @@ import {
 	PluginSidebar,
 	PluginSidebarMoreMenuItem,
 	PluginPrePublishPanel,
+	PluginPostPublishPanel,
 } from '@wordpress/edit-post';
 import { store as editorStore, PostTypeSupportCheck } from '@wordpress/editor';
 import { useState, useCallback } from '@wordpress/element';
@@ -42,13 +47,29 @@ registerPlugin( 'jetpack-social', {
 
 const JetpackSocialSidebar = () => {
 	const [ isModalOpened, setIsModalOpened ] = useState( false );
+	const [ isReviewRequestDismissed, setIsReviewRequestDismissed ] = useState(
+		getJetpackData()?.social?.reviewRequestDismissed ?? true
+	);
+	const [ shouldReviewRequestShow, setShouldReviewRequestShow ] = useState( false );
+
 	const openModal = useCallback( () => setIsModalOpened( true ), [] );
 	const closeModal = useCallback( () => setIsModalOpened( false ), [] );
 
 	const { hasConnections, hasEnabledConnections } = useSocialMediaConnections();
-	const { isPublicizeEnabled, hidePublicizeFeature, isSocialImageGeneratorEnabled } =
-		usePublicizeConfig();
+	const {
+		isPublicizeEnabled,
+		hidePublicizeFeature,
+		isPostAlreadyShared,
+		isSocialImageGeneratorEnabled,
+	} = usePublicizeConfig();
 	const isPostPublished = useSelect( select => select( editorStore ).isCurrentPostPublished(), [] );
+	// Determine if the review request should show right before the post publishes
+	// The publicize-enabled meta and related connections are disabled after publishing
+	usePostStartedPublishing( () => {
+		setShouldReviewRequestShow(
+			! isPostAlreadyShared && isPublicizeEnabled && hasEnabledConnections
+		);
+	}, [ isPostAlreadyShared, hasEnabledConnections, isPublicizeEnabled ] );
 
 	const PanelDescription = () => (
 		<Description
@@ -61,6 +82,22 @@ const JetpackSocialSidebar = () => {
 			} }
 		/>
 	);
+
+	// Handle when the review request is dismissed
+	const handleReviewDismiss = useCallback( () => {
+		const reviewRequestDismissUpdatePath =
+			getJetpackData()?.social?.dismissReviewRequestPath ?? null;
+		// Save that the user has dismissed this by calling to the social plugin API method
+		apiFetch( {
+			path: reviewRequestDismissUpdatePath,
+			method: 'POST',
+			data: { dismissed: true },
+		} ).catch( error => {
+			throw error;
+		} );
+
+		setIsReviewRequestDismissed( true );
+	}, [] );
 
 	return (
 		<PostTypeSupportCheck supportKeys="publicize">
@@ -107,6 +144,15 @@ const JetpackSocialSidebar = () => {
 			>
 				<SocialPreviewsPanel openModal={ openModal } />
 			</PluginPrePublishPanel>
+
+			{ ! isReviewRequestDismissed && shouldReviewRequestShow && (
+				<PluginPostPublishPanel id="publicize-title">
+					<ReviewPrompt
+						href={ getRedirectUrl( 'jetpack-social-plugin-reviews' ) }
+						onClose={ handleReviewDismiss }
+					/>
+				</PluginPostPublishPanel>
+			) }
 		</PostTypeSupportCheck>
 	);
 };
